Interim; not built
Created a timing generator and testbench.
The purpose of the timing generator is to create two separate "clock" signals for the cpu
One for the first half of the microcode cycle (when B bus is driven)
One for the second half when registers are captured on the C Bus.

The wave forms in the test bench look OK.
The B Bus wave form is symetric
And, for the C Bus, we generate a single pulse.

The key idea is everything is synchronized to the actual "clk".