On branch 20150610_mem_mapped_fsm
WOW!  Turns out problem writing to mem mapped device WAS
clock running at 100MHz!!!  ISE reports max speed is ~75Mhz.
Big change was to divide clock by 2, so my_clock is now 50Mhz.

Found this because STORE instruction worked but STORE2 instruction did NOT.
STORE2 did not work because it was using MDR which was not being set properly
because timing constraint was not being met.

There is some extra junk in this check in, but it does build and work with
instruction set tester (still not checked in).

Also checking in default.wcfg for use with simulator.
BTW Simulation looked good, b/c it did not capture timing problem.

The other reason I found problem was I found reads clearly worked and
writes with fixed values worked so problem was clearly with capturing data_bus
which was being driven by MDR.  (Putting in some dummy FSM states in write FSM
made problem go away, indicating data_bus was not stable.)