On branch 20150708_coregen_sim_mem Good interim checkin; Builds and runs asm test program
Goal was to create a "ram" using the coregen function in ISE.
Created a 1K * 8 ram and embedded it into fifo_test.vhd (which is
just an sram compatible entity).

Was able to simulate the ram, but not the "fifo" b/c the test bench
was not created properly.

Had to add blk_mem_gen_v7_3.ngc (from the ./ipcore_dir directory)
in order to build system.bit.

When run, read of 0xF040 returns 0x00AF - the default specified when
ram was created.

I tried the coregen, partly for the experience, partly to have an easier
entity.  Seems to have worked...