On branch 20150708_coregen_sim_mem
Builds and runs instruction set demo.
writes to "fifo" had previously worked but reads did not work for
addresses coming from address bus. (as opposed to a fixed addr which
seemed to be ok)

Had to delay by one clock cycle the capturing of the ram output.
(ie. the read fsm has an extra cycle)

I think the address is only captured on the rising edge of the clock
which would occur as fsm enters state 1 so val can only be captured
when LEAVING state 1.