On branch 20150710_pchu_uart_and_coregen_fifo Shocking interim checkin! (In a good way).
Builds and runs demo program.

PCHU RX is connected to Jamet coregen mem fifo and appears
to work.  Added a delay (in asm prog) and typed a bunch of chars
all of which were captured in ram and then printed.  I do not know if
wrapping logic works yet.

Also there are no status registers yet.