On branch 20150710_pchu_uart_and_coregen_fifo
Woowee! uart_w_fifo seems to work with both transmit and receive!
Ran a bunch of tests though not yet exhaustive.  Still this is a good interim
checkin.

The big thing to note is the extra state required when writing to the tx fifo.
Without this, the data is read by the serial transmitter before it is loaded.