On branch 20150710_pchu_uart_and_coregen_fifo Good interim checkin. This version of the cpu builds and runs the instruction set tester program including the pong chu uart with the large fifo's.
The idea is to have a bunch of status bits which indicate
empty, full, half and quarter states

The bits are now output from entity and (hopefully) can be used to generate
interrupts.

This is all part of the work to create a uart which matches the fifo uart in the
simulator.  The simulator assumed a separate uart just to indicate done conditions.
I hope to replace the second uart with one with approporate status output bits.