On branch 20150801_reconcile_vhdl_sim
Nice - This is the checkin with 3 of the new uarts_w_fifo on fpga
They are all mapped to PMOD D on the board and I made a custom cable
which goes to a custom board which connections to 3 Ada Fruit FTDI Friends.
All works exactly as expected.
Tested all ports on Windows and in LInux.
In Linux I used kermit (see the 3 kermit config files for kermit params)

These ports are consistent with the simulator, function and mem mapping.