Another misc check in.  This has to do the simulator. Turns out, running Pat's loader didn't work in simulator. It didn't work because of the memory protection features of the sim. This commit adds a feature to disable mem protection.
It also fixes a timing error with the console serial port which was running
about 10 times too fast.

Along the way I fixed a bug in the disk_ctlr.  Just a syntax error.
In fact this branch was (and still is) all about getting the disk
controller working with real serial ports.

Modified the second stage loader too.  It was working as expected, but
I changed the "halt" behaviour to spin in place so any error messages
would make it to a simulated console.  Previously a halt in the loader
stopped the simulator.

With this commit, I have seen the second stage loader work and saw it
load the kernel.