Interim but moving in the right direction. The Kernel can now load hex.sim files (0000 0002 magic) from the kernel cmd prompt.  Had to support memory protection types. Changed loader to use latest diskcontroller open read and close operations.
This is an interim step towards loading file system and shells which FIRST
requires the kernel to load procs and then lets us run those procs.

For example, we want to load the FS, run it and see it work with the
disk controller.  We've already seen this in sim; want to see it with
serial based disk_controller access.