On branch 20150819_pty_int_ctlr_disk_integration This is the last commit for this branch before merge. Only change was to big fifo uart status. address 1 was compatible with the uart I used for the first 9 years.  It had 2 status bits.  One indicated data received; the other indicated transmitter empty.
The transmitter empty bit was used by a lot of code,
including k_userio to indicate when a byte could be written.

Now that there's a 1KB fifo, I've changed that bit to indicate
tx NOT FULL.  This way no code changes are required and short
transmissions don't have to block while individual chars are transmitted.

As for state of system, now that the tctlr serial port on fpga is connected
directly to serial port on raspberry pi, the tctlr problems are gone.

It's possible to run 2 shells, 2 built-in shells, open close files, display
fib numbers all day long and get good responsiveness from the partial-cooked
terminal controller.

Seems very stable, but turning on my fluorescent task lamp still causes a system
crash in a VERY specific way. It causes the pty driver to request 200 chars from
the tctlr which hangs the system.

As of this writing, the other 2 ports (console and disk controller) both
still pass through serial<->usb adapters with no obvious problems.