Another very nice check in. 6 User procs running at once.
Bumped the number of tctlr buffers to 8 to match number of potential terminals (for now)
With 6 user procs running, 4 of which were the demo producing constant fib output,
I was unable to crash the system due to buffer overflow. I did this because I saw
it was possible to crash system if there was too much terminal i/o causing buffer exhaustion.

Modifiedl kernel to explicitly halt when num buffers is exceeded.

I think, as of this checkin, that the exclusive handling of the num_tctlr_buffers variable
probably needs to be rethought.