This is a trickier than expected target.  Here is where things stand.
I've been trying to create programs with split 64K Code / 64K Data.
Previous "commits" have incorporated changes to pcomp and jamasm and
a new file format was created called V3.  V3 contained enough info
to load two different banks.

Turns out the kernel cannot be split right now.  It's because the
K_SP_STORE instruction sets DS to 0x0000.

So to compensate for kernel problem, I created a V4 file.  V4 also splits
code and data but uses non-overlapping addresses so everything can be
placed in a single bank (as has always been the case with the kernel).

Well, in the last commit, the kernel would ultimately crash.  The sim
history showed code from pty driver would return to lower memory
and ultimately try to execute an invalid opcode.

I made a small change in this commit to pcomp because "guard" words
were being placed in code space. Now they are part of UDATA.

The change above "solved" the invalid opcode problem.  That's the good news.
BUT there are still cases where PTY messages cannot be delivered to
the PTY driver.  I do NOT believe this is related to the 64/64/ split.
I think this is related to a race condition involving pty buffer management.

I've seen the message delivery problem before which is why I do not think
the problem is 64/64/ related.  STILL THIS IS A GUESS.