Please read carefully! This is a no-change commit from a detached head branch based on 	Date:   Sat Feb 6 19:50:29 2016 -0500 	911186a70e107530984b8a5946527b8b1c41d891
It is not at the head of the current development as of this writing on
February 15th.

This commit is being made so I can note there is a problem with terminal i/o.

The full system runs for a while and then stops working properly.  In this
particular commit, I ran the 2 external shells and after an extended period
printing fibonacci numbers, the terminal system hung.  Since the entire system
hung (not just one process) it is probably the case that the pty driver got
input from the PTC with an incorrect  terminal number. The pty driver then
tries to send a message to a non existent process and hangs.

In later commits, I'd removed a lot of pty driver debugging messages AND
added sequence numbers to the PTC protocol.  Then the problems went away.
I DONT KNOW WHY!!!

So the purpose of this commit and branch is to figure out what's causing
the problem.

Have not been able to reproduce the problem in the simulator.  So either
the problem is with the h/w OR the simulator isn't accurate enough.  The hanging
problems occur in both the old Nexys board and the new arty board.