Upgraded pyserial to pyserial v3 and changed some method calls in serial_socket_shim to match. Also modified ptc to display sequence numbers.
As of this commit, the sequence ALWAYS look wrong in the ptc!!!
They should start at zero but consistently start at 0x48 instead.

Also the PTC hangs during the second host to term transmission.  I think
the reason is the count is incorrect and the PTC is hung waiting for data.

I am commiting this version b/c I believe the problems are communications related,
possibly due to lost serial characters going from host to ptc.  It's part of the reason
I upgraded pyserial (which had no effect...)