POSSIBLY encouraging commit...
Changed shim so it buffers ALL serial input data from fpga host before
sending it so socket connected PTC.
Had to make shim understand host->ptc protocol so it knew how much data to receive.

My thinking was/is that the serial data is arriving too fast for Linux to
read it serially and write it to socket.

REMOVED flush in socket to serial.  My thinking here is the serial library
will NOT allow me to overwrite its buffer.  This is different from the serial_to_socket
case because (in this case) the code is controlling the writes to the serial port.  In the
serial_to_socket case the serial data is arriving asynchronously from fpga host.

It's also important to note here that I'm now running the shim with pypy not python
so I will have to figure out (in future tests) if buffering is really the problem and if
just switching to pypy was/is sufficient.  THIS IS UNKNOWN AT THIS COMMIT.

As of this commit, the 2 shell constant output test ran once for a few hours.  Now,
I'm running it again...

All this is being done on the Nexys Spartan 3 board.