Some good news PTY; some bad news FS!
This version does not seem to hang in pty.
Ran n * 10k fib's on multiple terminals without a hang.
This MAY mean that the combination of intctlr fsm fix and DI/EI fix in pty
may have resulted in a stable system.

Now for the bad news.  The file i/o does not appear correct.  I do NOT know if this
is a new problem resulting from the work above OR if it's been latent.
When creating a new file and repeatedly writing to it, the size seems to jump
e.g. 0, 240, 480, 1200!!!
Another way to demo a problem is to write repeatedly and then lseek and write.
Later blocks do not appear correct.

Added missing default state to interrupt controller in host write fsm
kernel - re-enabled loading of multiple user procs
pty - disabled interrupts during entire code portion when tctlr buffers are being read
build_program - updated to reflect proper include dir